end2end_pipe/pyproject.toml
[build-system]
requires = ["setuptools>=67", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ml-portfolio-churn"
version = "0.1.0"
description = "Churn prediction CLI"
authors = [{name = "Mikkel Nielsen"}]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "numpy>=1.20.0",
  "pandas>=1.3.0",
  "scikit-learn>=1.0.0",
  "matplotlib>=3.4.0",
  "seaborn>=0.11.0",
  "joblib>=1.1.0",
  "liac-arff>=2.5.0",
  "pydantic>=2.6.0",
  "pandera>=0.19.0",
]

[project.optional-dependencies]
mlflow = [
  "mlflow>=2.12.0",
]
service = [
  "fastapi>=0.110.0",
  "uvicorn>=0.28.0",
  "prometheus-client>=0.20.0",
]

[project.urls]
Repository = "https://github.com/Medesen/ML-portfolio"

[project.scripts]
churn = "ml_portfolio_churn.main:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
include = ["ml_portfolio_churn*"]

